JRUBY-5125: let Cipher#name for 'rc4' to be 'RC4'.
Given 'rc4' by Cipher.new('rc4') must be 'RC4' when Cipher#name is
called.  OpenSSL checks it as a LN of object ID and set SN. We don't
check 'name' is allowed as a LN in ASN.1 for the possibility of JCE
specific algorithm so just do upperCase here for OpenSSL compatibility.