PKCS7 structures that have their content encapsulated in a sequence are not imported correctly
The case to handle content which has been encapsulated in a sequence
has been in place already but wasn't implemented. The fix is inspired
by Pull Request: https://github.com/jruby/jruby-ossl/pull/9. The test
case includes PKCS7 samples to showcase the issue. Credit goes to Jason
Franklin [https://github.com/jamuc] for helping with the samples and
the fix in general.