JRUBY-6961 Make openssl selectively load gem only if jruby < 1.7
This is a bit of a hack, but hopefully offers a better transition
experience to end users than the alternative of doing nothing.

With this change the included openssl.rb detects if it is running
under jruby 1.7+, and if so, loads what jruby's openssl.rb would load
instead of the gem contents.